ingestion-service/requirements.txt
requests
playwright
pinecone-client
sentence-transformers